requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
authentication==1.1.0
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==3.2.5
django-allauth==0.39.1
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.4
Flask==2.2.2
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.1
packaging==21.3
passlib==1.7.4
Pillow==9.2.0
protobuf==4.21.7
psycopg2==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.15.0
PyJWT==2.5.0
pyparsing==3.0.9
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.4
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.3
uritemplate==3.0.1
urllib3==1.26.12
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.1
